UI Changes
imported bootstrap for styling since it would be good for users accessing the site via there phone which is most.

imported Jquery/Javascript for if we need to add front end functionality to the webpage since im sure we will with the map.

improved the layout of the webpage and added a new chart.