Big changes
mpld3 was  not working properly and could not edit the layout for the graphs so i have changed the html converter to plotly which makes the map and graphs look way better.

I have added all the emissions data files to the project.

I added in a slider for selecting the year which when changing the year on the slider dynamically updates the data on the map and charts to the data for that year.

Follow the readme.txt to run the website.